+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 1.0.2
 
 - Added sorting in the dashboard
+- Instant set current speed
 - Fixed double-channel issue in the learning engine
+- Small improvements to the learning engine
 - Fixed small issues
 
 ## 1.0.1